fix: sequence breaker ids are a list of int

+++ aphrodite/common/sampling_params.py
@@ -219,7 +219,7 @@ class SamplingParams(
     dry_multiplier: float = 0.0
     dry_base: float = 1.75
     dry_allowed_length: int = 2
-    dry_sequence_breaker_ids: List[float] = []
+    dry_sequence_breaker_ids: List[int] = []
     # The below fields are not supposed to be used as an input.
     # They are set in post_init.
     output_text_buffer_length: int = 0
